requirements.txt
certifi==2017.4.17
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.3
django-allauth==0.32.0
django-filter==1.0.4
django-rest-auth==0.9.1
djangorestframework==3.6.3
drf-nested-routers==0.90.0
idna==2.5
oauthlib==2.0.2
python3-openid==3.1.0
pytz==2017.2
requests==2.18.1
requests-oauthlib==0.8.0
six==1.10.0
urllib3==1.21.1
